added data.photo for deleting image from firebase

+++ src/components/Account.jsx
@@ -164,11 +164,12 @@ const Account = ({ Loader, Error, ImageLoader }) => {
     }
     return;
   };
-  console.log(data.photo);
+  console.log(data?.photo);
   // delete user profile image
   const deleteProfilePic = async (e) => {
+    if (data?.photo === "default.jpg") return;
     setBtnProcessing(true);
-    console.log("imageURL for deleting", data.photo);
+    console.log("imageURL for deleting", data?.photo);
     const imageRef = ref(storage, imageURL);
     deleteObject(imageRef).then(() => {
       console.log("image delete successfully!");
